requirements.txt
python-docx>=0.8.11
jsonschema>=4.17.0
